backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "dotenv>=0.9.9",
    "fastapi>=0.119.1",
    "google-genai>=1.46.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
